WT-18068 Carry the local drop in the configuration
A layered table is reached through its column group, so the drop re-enters from
the top for each constituent and a parameter would have to be threaded through
every drop type. The configuration already travels that path, so read the local
drop from there and drop the session flag.

That returns the session flag count to what it was, so the flags field goes
back to 32 bits and the widening this change carried is no longer needed.

+++ src/checkpoint/checkpoint_conn.c
@@ -133,7 +133,7 @@ static int
 __ckpt_server_start(WT_CONNECTION_IMPL *conn)
 {
     WT_SESSION_IMPL *session;
-    uint64_t session_flags;
+    uint32_t session_flags;
 
     /* Nothing to do if the server is already running. */
     if (conn->ckpt.server.session != NULL)
