+++ src/conn/conn_sweep.c
@@ -630,7 +630,7 @@ int
 __wti_sweep_create(WT_SESSION_IMPL *session)
 {
     WT_CONNECTION_IMPL *conn;
-    uint64_t session_flags;
+    uint32_t session_flags;
 
     conn = S2C(session);
